+++ oshdb-util/src/main/java/org/heigit/bigspatialdata/oshdb/util/time/OSHDBTimestamps.java
@@ -156,6 +156,8 @@ public class OSHDBTimestamps implements OSHDBTimestampList {
 
       Period period = (Period) steps.get("period");
       Duration duration = (Duration) steps.get("duration");
+      
+      int counter = 1;
 
       //validate start and end. start should be before end.
       if (start.isAfter(end)) {
@@ -171,7 +173,8 @@ public class OSHDBTimestamps implements OSHDBTimestampList {
 
         while (currentTimestamp.toEpochSecond() >= startTimestamp) {
           timestamps.add(currentTimestamp.toEpochSecond());
-          currentTimestamp = currentTimestamp.minus(period).minus(duration);
+          currentTimestamp = end.minus(period.multipliedBy(counter)).minus(duration.multipliedBy(counter));
+          counter++;
         }
       } else {
         ZonedDateTime currentTimestamp = ZonedDateTime.from(start);
@@ -179,7 +182,8 @@ public class OSHDBTimestamps implements OSHDBTimestampList {
 
         while (currentTimestamp.toEpochSecond() <= endTimestamp) {
           timestamps.add(currentTimestamp.toEpochSecond());
-          currentTimestamp = currentTimestamp.plus(period).plus(duration);
+          currentTimestamp = start.plus(period.multipliedBy(counter)).plus(duration.multipliedBy(counter));
+          counter++;
         }
       }