+++ oshdb/src/test/java/org/heigit/bigspatialdata/oshdb/osm/OSMNodeTest.java
@@ -6,11 +6,6 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 import org.junit.Test;
 
-/**
- *
- * @author Rafael Troilo &lt;rafael.troilo@uni-heidelberg.de&gt;
- * @author Moritz Schott &lt;m.schott@stud.uni-heidelberg.de&gt;
- */
 public class OSMNodeTest {
 
   public OSMNodeTest() {}