+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+## 0.5.5
+
+* improved performance of data [stream](https://docs.ohsome.org/java/oshdb/0.5.4/oshdb-api/org/heigit/bigspatialdata/oshdb/api/mapreducer/MapReducer.html#stream--)ing queries on ignite (using AffinityCall backend).
+* make monthly time intervals more intuitive to use. #201
+* (minor) reorganize parent package: bigspatialdata-parent version bump to 1.2, rename bigspatialdata-core-parent to oshdb-parent
+
 ## 0.5.4
 
 * fix a regression where broken referential integrity in OSM data causes a crash during geometry building